Fixed signed/unsigned comparison warning that I introduced.
=============================
Build/Test Cases Summary
=============================

Wed Jan 20 15:13:38 EST 2010

Enabled Packages: Kokkos, Tpetra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=402,notpassed=0
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=426,notpassed=0

Local commits for this build/test group:
----------------------------------------
7843647 Fixed signed/unsigned comparison warning that I introduced.
f68b9a1 Multiply Hybrid node changes and corrections.
ff15774 Modified CrsMatrix to initially cache views upon matrix construction, to improve matrix fill on GPUs.
0c7a5a5 Disabled Kokkos::ThrustGPUNode on MPI tests; conflict over singular GPU resource causes failure.
2a1dd74 Added missing explicit instantiation for new Chebyshev method.

+++ packages/tpetra/example/PowerMethod/Tpetra_Power_Method_From_File.cpp
@@ -138,7 +138,7 @@ Scalar power_method(const Teuchos::RCP<const Tpetra::Operator<Scalar,Ordinal> > 
   Scalar lambda = static_cast<Scalar>(0.0);
   Magnitude normz, residual = static_cast<Magnitude>(0.0);
   // power iteration
-  for (size_t iter = 0; iter < niters; ++iter) {
+  for (int iter = 0; iter < niters; ++iter) {
     normz = z.norm2();                            // Compute 2-norm of z
     q.scale(ONE/normz, z);                        // Set q = z / normz
     A->apply(q, z);                               // Compute z = A*q
